remove Kerberos export
